BaboonFeed-B/requirements.txt
django
djangorestframework
django-channels
pillow
model-bakery
pytest-django
faker
ipython
django-rq
prettyconf
psycopg2